+++ src/components/index.js
@@ -1,9 +1,9 @@
-export { default as Button } from '@components/Button/Button.jsx';
-export { default as TabButton } from '@components/Button/TabButton.jsx';
+export { default as Button } from '@components/button/Button.jsx';
+export { default as TabButton } from '@components/button/TabButton.jsx';
 
-export { default as SearchBar } from '@components/SearhBar/SearchBar.jsx';
-export { default as CheckBox } from '@components/CheckBox/CheckBox.jsx';
-export { default as DropDown } from '@components/DropDown/DropDown.jsx';
+export { default as SearchBar } from '@components/searchBar/SearchBar.jsx';
+export { default as CheckBox } from '@components/checkBox/CheckBox.jsx';
+export { default as DropDown } from '@components/dropDown/DropDown.jsx';
 
 export { default as Card } from '@pages/common/event&policyRec/components/Card/Card.jsx';
-export { default as Paginations } from '@components/Pagination/Paginations.jsx';
+export { default as Paginations } from '@components/pagination/Paginations.jsx';
